Carbon v0.0.1 Carbon.Duration

Functions

format_minutes(duration_in_minutes)

Format a duration in minutes to a human readble string. Returns a string in the Format “1.5h”, “30m”, “2.5d”. There will never be more than one unit at the time. This function is the opposite of Carbon.Duration.parse_minutes/1

Examples

iex> Carbon.Duration.format_minutes(30)
"30m"

iex> Carbon.Duration.format_minutes(90)
"1h 30m"

iex> Carbon.Duration.format_minutes(420)
"1d"

parse_minutes(duration_string)

Parses a duration description. Returns the corresponding number of minutes as as integer. Duration descriptions are given in numeric value followed by a unit descriptor such as “m”, “h”, or “d”.

Examples

iex> Carbon.Duration.parse_minutes("30m")
30

iex> Carbon.Duration.parse_minutes("1h 30m")
90

iex> Carbon.Duration.parse_minutes("1d")
420