Carbon v0.0.1 Carbon.Router.Helpers

Module with named helpers generated from Carbon.Router.

Functions

account_activity_path(conn_or_endpoint, action, account_id)

account_activity_path(conn_or_endpoint, action, account_id, params)

account_activity_url(conn_or_endpoint, action, account_id)

account_activity_url(conn_or_endpoint, action, account_id, params)

account_contact_path(conn_or_endpoint, action, account_id)

account_contact_path(conn_or_endpoint, action, account_id, id)

account_contact_path(conn_or_endpoint, action, account_id, id, params)

account_contact_url(conn_or_endpoint, action, account_id)

account_contact_url(conn_or_endpoint, action, account_id, id)

account_contact_url(conn_or_endpoint, action, account_id, id, params)

account_deal_path(conn_or_endpoint, action, account_id)

account_deal_path(conn_or_endpoint, action, account_id, params)

account_deal_path(conn_or_endpoint, action, account_id, id, params)

account_deal_url(conn_or_endpoint, action, account_id)

account_deal_url(conn_or_endpoint, action, account_id, params)

account_deal_url(conn_or_endpoint, action, account_id, id, params)

account_event_path(conn_or_endpoint, action, account_id)

account_event_path(conn_or_endpoint, action, account_id, params)

account_event_path(conn_or_endpoint, action, account_id, id, params)

account_event_reminder_path(conn_or_endpoint, action, account_id, event_id)

account_event_reminder_path(conn_or_endpoint, action, account_id, event_id, params)

account_event_reminder_path(conn_or_endpoint, action, account_id, event_id, id, params)

account_event_reminder_url(conn_or_endpoint, action, account_id, event_id)

account_event_reminder_url(conn_or_endpoint, action, account_id, event_id, params)

account_event_reminder_url(conn_or_endpoint, action, account_id, event_id, id, params)

account_event_url(conn_or_endpoint, action, account_id)

account_event_url(conn_or_endpoint, action, account_id, params)

account_event_url(conn_or_endpoint, action, account_id, id, params)

account_path(conn_or_endpoint, action)

account_path(conn_or_endpoint, action, id)

account_path(conn_or_endpoint, action, id, params)

account_url(conn_or_endpoint, action)

account_url(conn_or_endpoint, action, id)

account_url(conn_or_endpoint, action, id, params)

instance_path(conn_or_endpoint, action)

instance_path(conn_or_endpoint, action, params)

instance_url(conn_or_endpoint, action)

instance_url(conn_or_endpoint, action, params)

page_path(conn_or_endpoint, action)

page_path(conn_or_endpoint, action, params)

page_url(conn_or_endpoint, action)

page_url(conn_or_endpoint, action, params)

path(data, path)

Generates the path information including any necessary prefix.

session_path(conn_or_endpoint, action)

session_path(conn_or_endpoint, action, params)

session_path(conn_or_endpoint, action, token, params)

session_url(conn_or_endpoint, action)

session_url(conn_or_endpoint, action, params)

session_url(conn_or_endpoint, action, token, params)

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

tag_path(conn_or_endpoint, action)

tag_path(conn_or_endpoint, action, params)

tag_path(conn_or_endpoint, action, id, params)

tag_url(conn_or_endpoint, action)

tag_url(conn_or_endpoint, action, params)

tag_url(conn_or_endpoint, action, id, params)

timesheet_path(conn_or_endpoint, action)

timesheet_path(conn_or_endpoint, action, id)

timesheet_path(conn_or_endpoint, action, id, params)

timesheet_timesheet_entry_path(conn_or_endpoint, action, timesheet_id)

timesheet_timesheet_entry_path(conn_or_endpoint, action, timesheet_id, params)

timesheet_timesheet_entry_path(conn_or_endpoint, action, timesheet_id, id, params)

timesheet_timesheet_entry_url(conn_or_endpoint, action, timesheet_id)

timesheet_timesheet_entry_url(conn_or_endpoint, action, timesheet_id, params)

timesheet_timesheet_entry_url(conn_or_endpoint, action, timesheet_id, id, params)

timesheet_url(conn_or_endpoint, action)

timesheet_url(conn_or_endpoint, action, id)

timesheet_url(conn_or_endpoint, action, id, params)

url(data)

Generates the connection/endpoint base URL without any path information.

user_path(conn_or_endpoint, action)

user_path(conn_or_endpoint, action, params)

user_path(conn_or_endpoint, action, id, params)

user_url(conn_or_endpoint, action)

user_url(conn_or_endpoint, action, params)

user_url(conn_or_endpoint, action, id, params)