Carbon v0.0.1 Carbon.SessionController

Functions

carbon_token_name()

create_and_send_session_link(conn, map)

Receives the email posted by the user.

If the email is known by the system, creates a short-lived token and sends a session link to the user. Otherwise, redirects the user to the account registration page passing along the email.

create_new_user(conn, map)

Receives the registration information posted by the user.

Creates a user and a short-lived token then sends a session link to the user. On User constraint error (duplicate handle or email) displays proper error messages.

index(conn, params)

Renders a login page where the user inputs her email.

logout(conn, params)

Deletes the session and redirects to the login page.

validate_token(conn, map)

Receives the request created by the session link.

If the token is still valid, finds the user associated to the token and assigns it to the web session; deletes the short-lived token; creates a long-lived token; adds a cookie to the response with the new long-lived token. Redirects the user to the landing page.

Otherwise, redirects the user to the login page with proper error message.